update function get_min.m
use parameters from the struct params that is defined in the file hyperParam.m. This struct is passed as argument, instead of using global variables.

correct H_mat (use H instead of H^2).

correct input to to the meshgrid function.

correct X_jt and Y_jt

add the factor to get correct result