update optimize_m_debug
argument list: deleted redundant and useless parameters. The function takes the hover points and the whole trajectory. From this the initial trajectory is extracted.

move the calculation of the hover points outside of this function.

use the new version of the crb functions.

dimension of S is now the same as V.

clean code: change name of variables to a meaningful and less confusing ones. Delete redundancy from names